Read TCNT2 only once in get_count()
The get_count() method would read TCNT2 a second time whenever an
overflow was detected. This can cause a slight jitter, as the time
between entering the method and taking the last reading of TCNT2 is not
constant. We fix this by reading it only once, at the start of the
method.

In order to ensure a correct result across an overflow, we examine the
value we got in order to know whether it was read right before or right
after the overflow.